Xiangshan volcano basin topography
The topography data were collected in the Xiangshan volcano basin (Fig. 6). The inclined 3D axial anisotropic body is embedded in an isotropic half-space with a resistivity of 100 ohm.m. The anisotropic body is 20 km in strike length, has a thickness of 400 m and dips 53°, and extends 10 km along the dipping direction.